Updated all tests to the new haskell compatible syntax.
There is some regression here, due to the use of more 'import's. The
problem is orthogonal to the new syntax: since I had to edit all the
files, I decided to use 'import's more effectively.  However, this
results in a many irrelevant definitions being in scope, and so in
turn the generated theory is much larger in some cases.

There are two obvious fixes:

1. go back to more granular files, with fewer or more granular
'import's

2. make the theory generation phase only generate the relevant axioms.

(1) is easier in the short run, but sloppy.  (2) is harder, but a good
approximation is probably straightforward: transitively search for
free variables, and generate axioms for all names found this way.

In detail, here's how the 20 second timeout tests degraded:

  $ diff -U0 results/46.20s-equinox.txt results/47.20s-equinox.txt
  --- results/46.20s-equinox.txt  2011-11-08 15:20:04.000000000 +0000
  +++ results/47.20s-equinox.txt  2011-11-15 18:52:11.000000000 +0000
  @@ -14,2 +14,2 @@
  -./yes/bintree-height-is-cf-to-cf.hs:              passed
  -./yes/bintree-size-is-cf-to-cf.hs:                passed
  +./yes/bintree-height-is-cf-to-cf.hs:              TIMED OUT
  +./yes/bintree-size-is-cf-to-cf.hs:                TIMED OUT
  @@ -17,0 +18 @@
  +./yes/gt-inc.hs:                                  TIMED OUT
  @@ -23 +24 @@
  -./yes/prediate-on-constant-function.hs:           passed
  +./yes/predicate-on-constant-function.hs:          FAILED WITH EXIT CODE 1
  @@ -30 +31 @@
  -./no/add-and-mult-nonZero.hs:                     FAILED WITH EXIT CODE 1
  +./no/add-and-mult-nonZero.hs:                     timed out
  @@ -33,0 +35 @@
  +./no/mult-gt.hs:                                  FAILED WITH EXIT CODE 1

The tests no/mult-gt.hs and yes/gt-inc.hs are new since results/46.