Added example illustrating that multiple contract translation bug.
I had previously noticed that multiple contracts on the same function
would produce the wrong results, but I just realized we have the same
problem when we have multiple contracts on mutually recursive functions.
In the translation we produce, we get a refutation if *any* of the
contracts hold, instead of only when they *all* hold.